pyproject.toml
[tool.poetry]
name = "ej-platform"
version = "3.8.2"
description = "A Django-powered participation platform"
authors = ["Fábio Macêdo Mendes <fabiomacedomendes@gmail.com>, Pencillabs <contato@pencillabs.tec.br>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = ">= 3.9.19, <3.10"
django = "4.2.0"
django-constance = {version = "^2.3", extras = ["database"]}
django-allauth = "0.59.0"
django-taggit = "3.0.0"
django-autoslug = "^1.9"
django-environ = "^0.4.5"
django-bulk-update = "^2.2"
django-manager-utils = "^3.0.1"
django-polymorphic = "^2.0"
django-pandas = "^0.5.1"
django-filter = "^1.1"
django-model-utils = {git = "https://github.com/jazzband/django-model-utils.git", rev = "8e19699b0f8edbb93863de5c72256017080f2af0"}
hyperpython = "^1.1.1"
rules = "<2.0"
sidekick = "==0.5.2"
scipy = "^1.2"
pandas = "^1.2"
scikit-learn = "^1.0"
matplotlib = "^3.0"
invoke = "2.0.0"
wordcloud = "^1.5"
stop-words = "^2018.7"
djangorestframework = "3.14.0"
django-cors-headers = "^3.3.0"
babel = "^2.9.0"
sphinx = "5.0.2"
psycopg2 = "==2.8.6"
sphinx-rtd-theme = "<2.0.0"
toml = "^0.10.0"
gunicorn = "^19.9"
jinja2 = "2.11.3"
django-debug-toolbar = "3.2.0"
setuptools = "61.2.0"
django-anymail = {extras = ["mailgun"], version = "^10.0"}
django-ckeditor = "6.3.2"
colorama = "^0.4.6"
django-boogie = {git = "https://gitlab.com/pencillabs/ej/django-boogie.git", rev = "fixes-python39"}
importlib-metadata = "^7.1.0"
argon2-cffi = "^23.1.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
numpy = "^1.26.4"
libsass = "^0.23.0"
drf-spectacular = "^0.27.2"
django-jazzmin = "^3.0.0"

[tool.poetry.dev-dependencies]
bs4 = "^0.0.1"
mock = "^2.0"
model_mommy = "^1.6"
pytest = "==7.0.0"
pytest-django = "4.9.0"
black = {version = "^22.3.0", allow-prereleases = true}
ruff = "0.3.7"
ipython = "^7.2"
watchdog = "^0.9.0"
pre-commit = "^1.17"
notebook = "^6.0"
django-debug-toolbar = "3.2.0"
faker = "^24.4.0"

[tool.black]
line-length = 90
target_version = ['py38']
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.ej.conf]
themes=["default"]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ej.settings"
django_find_project = true
pythonpath = "src"
testpaths = "src"
norecursedirs = [".tox", "lib"]
addopts = "--nomigrations"

[tool.ruff.lint.per-file-ignores]
"src/ej/components/__init__.py" = ["F401"]
"src/ej/testing/__init__.py" = ["F401"]
"src/ej_clusters/math/__init__.py" = ["F401"]
"src/ej_clusters/models/__init__.py" = ["F401"]
"src/ej_conversations/__init__.py" = ["F401"]
"src/ej_boards/__init__.py" = ["F401"]
"tasks.py" = ["F403"]
"src/ej/tests/__init__.py" = ["F403"]
"src/ej/roles/__init__.py" = ["F401","F403"]
"src/ej_conversations/roles/__init__.py" = ["F401","F403"]
"src/ej_conversations/models/__init__.py" = ["F401","F403"]
"src/ej_integrations/mailing.py" = ["F523"]
"src/ej_integrations/tests/conftest.py" = ["F403"]
"src/ej_clusters/tests/conftest.py" = ["F403"]
"src/ej_users/apps.py" = ["F401"]
